requirements.txt
flask
requests
redis
APScheduler
python-dotenv